Use new data for history graph color
Previously the “isUp” status was retrieved from the cursor in the main
activity and passed to the fragment. This value was based on a positive
or negative `change` field from the API, however the same value can be
calculated by looking at the close value on the last two history
datapoints, completely isolating it to the fragment. For the widget
workflow to open the fragment dialog from an intent, this is required.